updating to address feedback
