+++ Assets/HoloToolkit/Utilities/Scripts/RaycastResultHelper.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 7899fdad227825d4dad5c0992e76f250
+timeCreated: 1514912152
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
